fix: amount datatype

+++ lib/screens/add_transaction.dart
@@ -49,8 +49,8 @@ class _AddTransactionState extends State<AddTransaction> {
         (_titleController.isEmpty) ? "Adjusted Balance" : _titleController,
         _descriptionController,
         (_typeController.toString().toLowerCase() == "income")
-            ? int.parse('+$_amountController')
-            : int.parse('-$_amountController'),
+            ? double.parse('+$_amountController')
+            : double.parse('-$_amountController'),
         (_walletController.isEmpty) ? "cash" : _walletController,
         _typeController.toLowerCase(),
         _categoryController,
